+++ plugins/gs/src/apis/auth/overrides/oauth2/lib/AuthConnector/index.ts
@@ -1,0 +1,2 @@
+export { CustomAuthConnector } from './CustomAuthConnector';
+export * from './types';