+++ plugins/gs/src/apis/auth/overrides/oauth2/lib/AuthSessionManager/index.ts
@@ -1,0 +1,1 @@
+export { RefreshingAuthSessionManager } from './RefreshingAuthSessionManager';