+++ plugins/gs/src/components/hooks/useResourceRequests.ts
@@ -5,11 +5,11 @@ import {
   type ResourceRequest,
 } from '@giantswarm/backstage-plugin-gs-common';
 import { useApi } from '@backstage/core-plugin-api';
-import { kubernetesApiRef } from '@backstage/plugin-kubernetes';
 import { useQueries } from '@tanstack/react-query';
 import { getK8sGetPath } from './utils/k8sPath';
 import { getInstallationsQueriesInfo } from './utils/queries';
 import { useApiVersionOverrides } from './useApiVersionOverrides';
+import { gsKubernetesApiRef } from '../../apis/kubernetes';
 
 export function useResourceRequests(
   kratixResources: {
@@ -22,7 +22,7 @@ export function useResourceRequests(
   const apiVersionOverrides = useApiVersionOverrides(
     kratixResources.map(item => item.installationName),
   );
-  const kubernetesApi = useApi(kubernetesApiRef);
+  const kubernetesApi = useApi(gsKubernetesApiRef);
   const queries = useQueries({
     queries: kratixResources.map(
       ({ kind, name, namespace, installationName }) => {