+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # backend
 
+## 0.2.0
+
+### Minor Changes
+
+- 06092e9: Change GS auth backend module to support client side OIDC provider.
+- f508faf: Update Backstage packages to v1.32.5.
+
+### Patch Changes
+
+- 06092e9: Move custom GitHub auth provider from GS backend module to backend package.
+- Updated dependencies [06092e9]
+- Updated dependencies [f508faf]
+- Updated dependencies [06092e9]
+  - @giantswarm/backstage-plugin-auth-backend-module-gs@0.3.0
+  - @giantswarm/backstage-plugin-scaffolder-backend-module-gs@0.3.0
+  - @giantswarm/backstage-plugin-techdocs-backend-module-gs@0.3.0
+
 ## 0.1.0
 
 ### Minor Changes